requirements.txt
openai>=1.0.0
requests>=2.28.0
Pillow>=9.0.0
python-dotenv>=0.19.0
urllib3>=1.26.0
